backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-openai==1.1.9
langgraph==1.0.8
langsmith==0.7.3
neo4j==5.25.0
yfinance==0.2.48
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
